Created my main Project class and my first example class.
I went ahead and created a project class that I could use for all of
the example classes in the book. The other class,Dog.java, is the first
example I worked on just to create a simple dog that had a name, size,
and breed. The dog also would have a method bark.