Added the GuessGame and Player example classes
These two classes are designed to work together to make and run a number
guessing game. The game itself makes three players that are objects of
the Player class. This Player class has a method that allows each one
to guess a random number and then that be assigned to their one
attribute by the same name. From there the game picks a random number
and then checks to see if each of the three players got it right. If not
their variable, which starts off false, remains the same. If true,
however, then it is assigned to be true. From there the game announces
who won and then breaks out of the while loop. If no winner the process
is repeated until at least one is found.