Added Gladiator class that is not in the book
I have went in and added a Gladiator class that is fairly simple. Each
Gladiator object has a name, health, and strength. This class plus
another I will build called Battle will allow two Gladiators to fight
each other. I did this as a practice of what I've learned in the book
and with Java so far.