Added Battle class and added to Gladiator class
I've started working on Battle to see how the two Gladiator classes
will interact and fight each other. I also added more to the
Gladiator class as well as a variable call isDead which was a Boolean
that stated whether or not the Gladiator was dead or not. This will
later be determined by their health going below zero.